Correctly store references to tracks that correspond to sound modules
Tracks are accessed based on their index. To keep a reference to a track, we need its index in a TrackBank. For now, we will create a large TrackBank and keep track of how many tracks are in the project. We know the index of a new track is one less than the total number of tracks.

This approach will break for projects that have more than 64 tracks (the current size of the TrackBank).