Handle preset browsing using new BrowserItemSelector
The BrowserItemSelector is an abstraction for searching a browser column for an item with the given name.

In this case, it is used to select the correct device type (Instrument) when browsing. It is also used to select the correct device when browsing for presets so that only presets that apply to the selected device are shown.